Extract shared frame prefetch acceleration helpers
Add infra.acceleration.frame_prefetch with CUDA host prefetch, lazy CUDA-frame
materialization, and generic prefetch dispatch helpers. Migrate the OmniDreams
raster and world-model lazy frame paths to share the implementation while
keeping the old cuda_host_prefetch import path as a compatibility shim.
